requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0.1

numpy>=1.26

sentence-transformers>=3.0.1
transformers>=4.41

# Reranking
cross-encoder>=1.0.0

pymupdf>=1.24
python-docx>=1.1

openai>=1.30

# Web UI
streamlit>=1.34

# FAISS wheels are typically available on Linux; on Windows use Docker/WSL/Conda or the built-in NumPy fallback.
faiss-cpu; platform_system != "Windows"
